Fixed an issue with Click and Tempo
Fixed issue where Click wouldn't be internally consistent between click() and getBasis() when the internal period parameter was changed dynamically. This caused issues where the Tempo ramp() wouldn't be consistent with beat() when you changed the bpm.

Additionally, fixed an issue where Tempo wouldn't fire beat() on the same frame as a trigger(), leading to beat() being conceptually called at the end of the beat instead of the beginning.