cache broadcast addresses on Windows
Just enumerating network interfaces with
NetworkInterface.getNetworkInterfaces() on Windows can take up to 200ms,
and this method is called from a number of places that assume it will
return quickly. Instead of refactor every user of this method, this adds
a cache for returned data. The cache has some disadvantages with respect
to interactivity (it takes longer to discover new broadcast addresses),
so the cache is only used on Windows where this method is a problem.