calculate network frame rate using monotonic clock
currentTimeMillis is a bad choice for frame rate calculation
because it skews with the system's clock, so things like NTP
clock skew, the user changing the time zone or date, etc. can
mess up the frame rate calculation. Plus, using nanoTime gets
you a more accurate frame rate (not that that matters much).