fix memory leak by caching PFonts loaded from disk
There are some UI components that load a font by name, and make no effort
to load them staticly (which itself would be difficult). This wouldn't be a
huge issue, except PFonts are staticly cached on the Processing side,
which means they have an infinite lifetime.

The real problem is Processing's caching; it should either be returning
the cached font when we ask for it again, or not caching them at all. But since
LX is easy to change and Processing is pretty much impossible, this is the
layer at which I'm fixing this memory leak.