fix workspace switching to project 0
Previously, the Workspace was always initialized thinking it was
on project zero. This isn't always the case; it's totally possible
for SLStudio to start up with a project loaded that isn't the first
in the Workspace. Workspace would then refuse to switch to project
zero because it thought it was already in it.