better detection of JAVA_HOME (more info below)
The java.home property works well on Linux even when the user hasn't set
JAVA_HOME in the containing shell. But on MacOS + Homebrew-installed
JDK8, the java.home system property points to $JAVA_HOME/jre, which
probably is just reflecting the system's preference to run out of the
JRE instead of the JDK the JRE is installed inside.

The preferred way to get the value for JAVA_HOME on Mac is to run
/usr/libexec/java_home, but it's tricky to get gradle to fill in the
value of a variable from the output of a command line. Instead of
attempting to get that working, we just throw up a warning telling mac
users how to set JAVA_HOME in the containing shell.