switch to nonblocking writes on AllPixelOutput
Sometimes a write hangs forever, and the timeout doesn't work on
anything but Windows. Instead, we just best-effort write it out with no
blocking at all, and then try to read from the socket; if the read
fails (and the read timeout does work on all platforms) then we know we
need to reestablish our connection to the output.