use workaround for JVM bug JDK-8019496
startOnFirstThread is needed because MacOS requires that windows be
created from the main thread of a process. But some versions of the JDK
have a deadlock that occurs when (a) AWT is loaded and (b)
-XstartOnFirstThread is used and (c) you try to load native libraries
from the JDK distribution itself.

We need to use startOnFirstThread, and we need to use ImageIO, which is
a native library in the JDK distribution, so the only option for us is
to not use AWT.

The fun thing about this is that it breaks our file dialogs, because we
were using AWT for those! So now we need a different solution for that.
But at least this will prevent us from hanging the whole process when
you select one of the FlockWave patterns.