make LXMidiDevice.getDescription work as intended on Windows
Everything in LX/SLStudio assumes that the actual identifier of a MIDI
device is in the MIDI device's description, not the name. That works
nicely on Linux and MacOS, where CoreMIDI sets the name to "CoreMIDI4J -
$(a truncated name)" and the description to just the full name of the
controller.

Windows sticks "No details available" in the description field for
physical devices, which is useless other than to tell us that we're on
Windows and we should probably look at the name field instead. I don't
see a great solultion here, this is the best one I could come up with.